+++ src/utils/random/index.ts
@@ -92,3 +92,10 @@ export function randomString(length = 12, options: RandomStringOptions = {}) {
 
   return result;
 }
+
+/**
+ * 生成一个随机的 boolean 值
+ */
+export function randomBoolean() {
+  return Math.random() < 0.5;
+}